pyproject.toml
[tool.poetry]
name = "energy_availability_broker"
version = "0.1.0"
description = "Communication architecture demonstration for multi level demand-response"
authors = ["Andreas Clement <a.clement@ptw.tu-darmstadt.de>, Tobias Koch <t.koch@ptw.tu-darmstadt.de>"]
readme = "README.md"
license = "BSD-2-Clause License"

keywords = ["eta_x", "experiment", "demand_response", "multi_agent_system"]

# specify some classifiers, to signify, who you are developing for. See: https://pypi.org/classifiers/
classifiers = [
    "License :: OSI Approved :: BSD License",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Information Analysis"
]

# Specify URLs for the project (optional)
# homepage = "https://www.ptw.tu-darmstadt.de" # TODO: change this
# documentation = "https://eta-utility.readthedocs.io/" # TODO: change this
# repository = "https://github.com/PTW-TUDa/eta_utility/" # TODO: change this


[tool.poetry.dependencies]
python = ">=3.11.0,<3.12"
numpy = ">=2.0.0,<3.0.0"
fastapi = {extras = ["standard"], version="^0.115.8"}
anyio = "^4.8.0"
apscheduler = "4.0.0a5"
pandas = "^2.2"
xgboost = "^3.0"
joblib = "^1.4"
openmeteo-requests = "^1.6"
retry-requests = "^2.0"
requests-cache = "^1.2"
eta-nexus = {git="https://git.ptw.maschinenbau.tu-darmstadt.de/eta-fabrik/public/eta-nexus.git", rev="ae8427cdda238a3154bd6c8e297098cdc074fc8b"}
marshmallow = "<4.0.0"
py-base-types = {path = "vendor/wheels/py_base_types-4.1.3-py3-none-any.whl"}
etaone-py-sdk = {path = "vendor/wheels/etaone_py_sdk-4.1.8-cp311-cp311-win_amd64.whl"}


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.5.0"
ruff = "^0.15.9"
pytest = "^8.3.4"
pytest-mock = "^3.14.0"

[tool.poetry.group.notebook.dependencies]
requests = "^2.31"
pandas = "^2.2"
matplotlib = "^3.8"
notebook = "^7.0"
python-dotenv = "^1.0"
numpy = "^2.0"
seaborn = "^0.13"
scipy = "^1.11"

[tool.poetry.scripts]
retrain_model = "energy_availability_broker.retrain:main"
fastapi_dev = "energy_availability_broker.dev_cli:fastapi_dev"

[tool.ruff]
include = ["pyproject.toml", "energy_availability_broker/**/*.py", "test/**/*.py", "examples/**/*.py"]

line-length = 120
indent-width = 4

required-version = ">=0.4.0"
target-version= "py39"

# Only fix with $ruff check --fix
fix = false

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends", "fastapi.params.Depends", "fastapi.Query", "fastapi.params.Query"]

[tool.ruff.lint]
select = [
    "F",       # Pyflakes
    "E", "W",  # Pycodestyle, Error and Warning
    "I",       # Isort
    "N",       # Pep8-naming conventions
    "UP",      # Pyupgrade
    "PL",      # Pylint
    "ERA",     # Eradicate
    "PD",      # Pandas-vet
    "NPY",     # NumPy-specific
    "RUF",     # Ruff

    # Flake8:
    # asyncio, bugbear, builtins, commas, comprehensions, eradicate, future-annotations,
    # implicit-str-concat, logging, no-pep420, print, pyi, pytest-style, return, simplify, use-pathlib
    "ASYNC", "B", "A", "COM", "C4", "FA", "ISC", "LOG", "INP", "T20", "PYI", "PT", "RET", "SIM", "PTH",

    ]

ignore = [
    "ERA001", #commented code
    "SIM105",
    "PLR2004",  # Pylint: magic values
    # Recommended to ignore as they conflict with the formatter
    "W191", "E111", "E114", "E117", "D206", "D300", "COM812", "COM819", "ISC001"
]


[tool.ruff.lint.extend-per-file-ignores]
"__init__.py" = ["PLC0414"]
"examples/*" = ["T201"]
"test/*" = ["RUF012"]

[tool.ruff.lint.isort]
combine-as-imports = true
force-wrap-aliases = true

[tool.ruff.lint.pylint]
max-args = 17
max-branches = 20
max-statements = 64
max-returns = 9
max-nested-blocks = 20

[build-system]
requires = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"
